feat: implement authentication system with breeze scaffolding
Add Laravel Breeze package for authentication scaffolding
Create auth controllers, views, and routes
Implement email verification and password reset functionality
Add user roles (admin/regular) and seed test users
Update user model with role checks and relationships
Include comprehensive tests for all auth features

+++ database/migrations/2025_08_21_122105_create_events_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->dateTime('start_date')->nullable();
             $table->dateTime('end_date')->nullable();
             $table->integer('capacity')->nullable();
-            $table->unsignedInteger('ticket_price')->nullable();
+            $table->decimal('ticket_price', 10, 2)->nullable();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->timestamps();
         });
